proceso de spammer de actor

+++ Acme-HackerRank/src/main/webapp/views/message/list.jsp
@@ -28,7 +28,7 @@
 		</a>
 	</display:column>
 	<display:column>
-		<a href="message/administrator,company,hacker/delete.do?messageId=${row.id}">
+		<a href="message/administrator,company,hacker/delete.do?messageId=${row.id}" onclick="return confirm('<spring:message code="message.confirm.delete"/>')">
 			<spring:message code="message.delete" />
 		</a>
 	</display:column>
@@ -50,7 +50,7 @@
 		</a>
 	</display:column>
 	<display:column>
-		<a href="message/administrator,company,hacker/delete.do?messageId=${fila.id}">
+		<a href="message/administrator,company,hacker/delete.do?messageId=${fila.id}" onclick="return confirm('<spring:message code="message.confirm.delete"/>')">
 			<spring:message code="message.delete" />
 		</a>
 	</display:column>
